backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.40
aiosqlite==0.21.0
opencv-python==4.11.0.86
mediapipe==0.10.21
numpy==1.26.4
python-dotenv==1.1.0
pydantic-settings==2.11.0
email-validator==2.3.0
